Refactor components for es6+ compile and circular errors (#3151)
* refactor(es6): resolve es6+ compile errors and most circular deps #3011

Use base class or a token/interface combo for DI child components to
avoid Circular dependency warnings and compilation errors in targets
es2015 and newer. Added explicit constructors in extending classes as
DI metadata doesn't get added otherwise (subject to fixes in ng compiler)

Also resolved some general circular refs and updated
tests that wouldn't work with es6.

Affected: list, calendar, combo, input-group, drop-down, expansion-panel,
grids, tabs, time-picker

* fix(hammer-configs): Add  Injectable DatePicker/TimePicker HammerConfigs. #3011

(cherry picked from commit 59aed5e5bce40d5cdfc68c904baa2987ab795190)
