+++ projects/igniteui-angular/src/lib/calendar/index.ts
@@ -1,4 +1,4 @@
 export * from './calendar';
-export * from './calendar.directives';
 export * from './calendar.component';
+export * from './calendar.directives';
 export * from './calendar.module';
