+++ projects/igniteui-angular/src/lib/drop-down/drop-down.common.ts
@@ -1,0 +1,17 @@
+import { CancelableEventArgs } from '../core/utils';
+import { IgxDropDownItemBase } from './drop-down.base';
+
+/** @hidden */
+export enum Navigate {
+    Up = -1,
+    Down = 1
+}
+
+/**
+ * Interface that encapsulates onSelection event arguments - old selection, new selection and cancel selection.
+ * @export
+ */
+export interface ISelectionEventArgs extends CancelableEventArgs {
+    oldSelection: IgxDropDownItemBase;
+    newSelection: IgxDropDownItemBase;
+}
