+++ projects/igniteui-angular/src/lib/expansion-panel/index.ts
@@ -1,5 +1,5 @@
-
-export { IgxExpansionPanelHeaderComponent, IExpansionPanelEventArgs } from './expansion-panel-header.component';
+export { IExpansionPanelEventArgs } from './expansion-panel.common';
+export { IgxExpansionPanelHeaderComponent } from './expansion-panel-header.component';
 export { IgxExpansionPanelBodyComponent } from './expansion-panel-body.component';
 export { IgxExpansionPanelComponent } from './expansion-panel.component';
 export * from './expansion-panel.module';
