+++ projects/igniteui-angular/src/lib/grids/column.component.ts
@@ -1372,4 +1372,9 @@ export class IgxColumnGroupComponent extends IgxColumnComponent implements After
     }
 
     set width(val) { }
+
+    constructor(public gridAPI: GridBaseAPIService<IgxGridBaseComponent>, public cdr: ChangeDetectorRef) {
+        // D.P. constructor duplication due to es6 compilation, might be obsolete in the future
+        super(gridAPI, cdr);
+    }
 }
