+++ projects/igniteui-angular/src/lib/grids/tree-grid/tree-grid-row.component.ts
@@ -1,9 +1,11 @@
-import { Component, forwardRef, Input, ViewChildren, QueryList, HostBinding } from '@angular/core';
+import { Component, forwardRef, Input, ViewChildren, QueryList, HostBinding, ElementRef, ChangeDetectorRef } from '@angular/core';
 import { IgxTreeGridComponent } from './tree-grid.component';
 import { IgxRowComponent } from '../row.component';
 import { ITreeGridRecord } from './tree-grid.interfaces';
 import { IgxTreeGridAPIService } from './tree-grid-api.service';
 import { State, TransactionType } from '../../services';
+import { GridBaseAPIService } from '../api.service';
+import { IgxSelectionAPIService } from '../../core/selection';
 
 @Component({
     selector: 'igx-tree-grid-row',
@@ -11,6 +13,14 @@ import { State, TransactionType } from '../../services';
     providers: [{ provide: IgxRowComponent, useExisting: forwardRef(() => IgxTreeGridRowComponent) }]
 })
 export class IgxTreeGridRowComponent extends IgxRowComponent<IgxTreeGridComponent> {
+    constructor(
+        public gridAPI: GridBaseAPIService<IgxTreeGridComponent>,
+        selection: IgxSelectionAPIService,
+        public element: ElementRef,
+        public cdr: ChangeDetectorRef) {
+            // D.P. constructor duplication due to es6 compilation, might be obsolete in the future
+        super(gridAPI, selection, element, cdr);
+    }
     private _treeRow: ITreeGridRecord;
 
     /**
