+++ projects/igniteui-angular/src/lib/input-group/input-group.common.ts
@@ -1,0 +1,7 @@
+/** @hidden */
+export abstract class IgxInputGroupBase {
+    disabled: boolean;
+    isFocused: boolean;
+    isRequired: boolean;
+    hasPlaceholder: boolean;
+}
