+++ projects/igniteui-angular/src/lib/list/index.ts
@@ -1,3 +1,3 @@
 export * from './list.component';
-export * from './list.common';
+export * from './list.common'; // TODO: all exports?
 export * from './list-item.component';
