+++ projects/igniteui-angular/src/lib/list/list.common.ts
@@ -1,7 +1,21 @@
-import { Directive, TemplateRef } from '@angular/core';
+import { Directive, TemplateRef, EventEmitter, QueryList } from '@angular/core';
 
 export interface IListChild {
     index: number;
+}
+
+/** @hidden */
+export abstract class IgxListBase {
+    onItemClicked: EventEmitter<any>;
+    allowLeftPanning: boolean;
+    allowRightPanning: boolean;
+    panEndTriggeringThreshold: number;
+    onLeftPan: EventEmitter<any>;
+    onRightPan: EventEmitter<any>;
+    onPanStateChange: EventEmitter<any>;
+    children: QueryList<any>;
+    listItemLeftPanningTemplate: IgxListItemLeftPanningTemplateDirective;
+    listItemRightPanningTemplate: IgxListItemRightPanningTemplateDirective;
 }
 
 export enum IgxListPanState { NONE, LEFT, RIGHT }
