+++ projects/igniteui-angular/src/lib/time-picker/time-picker.common.ts
@@ -1,0 +1,26 @@
+import { ElementRef } from '@angular/core';
+
+/** @hidden */
+export const IGX_TIME_PICKER_COMPONENT = 'IgxTimePickerComponentToken';
+
+/** @hidden */
+export interface IgxTimePickerBase {
+    _ampmItems: any[];
+    hourList: ElementRef;
+    minuteList: ElementRef;
+    ampmList: ElementRef;
+    selectedHour: string;
+    selectedMinute: string;
+    selectedAmPm: string;
+    nextHour();
+    prevHour();
+    nextMinute();
+    prevMinute();
+    nextAmPm();
+    prevAmPm();
+    okButtonClick(): boolean;
+    cancelButtonClick(): void;
+    scrollHourIntoView(item: string): void;
+    scrollMinuteIntoView(item: string): void;
+    scrollAmPmIntoView(item: string): void;
+}
