+++ projects/igniteui-angular/src/lib/time-picker/time-picker.component.ts
@@ -16,7 +16,8 @@ import {
     ViewChild,
     AfterViewInit,
     DoCheck,
-    ContentChild
+    ContentChild,
+    Injectable
 } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { HAMMER_GESTURE_CONFIG, HammerGestureConfig } from '@angular/platform-browser';
@@ -33,8 +34,10 @@ import {
 } from './time-picker.directives';
 import { Subscription } from 'rxjs';
 import { EditorProvider } from '../core/edit-provider';
+import { IgxTimePickerBase, IGX_TIME_PICKER_COMPONENT } from './time-picker.common';
 
 let NEXT_ID = 0;
+@Injectable()
 export class TimePickerHammerConfig extends HammerGestureConfig {
     public overrides = {
         pan: { direction: Hammer.DIRECTION_VERTICAL, threshold: 1 }
@@ -62,13 +65,24 @@ export interface IgxTimePickerValidationFailedEventArgs {
         {
             provide: HAMMER_GESTURE_CONFIG,
             useClass: TimePickerHammerConfig
+        },
+        {
+            provide: IGX_TIME_PICKER_COMPONENT,
+            useExisting: IgxTimePickerComponent
         }
     ],
     selector: 'igx-time-picker',
     styles: [':host {display: block;}'],
     templateUrl: 'time-picker.component.html'
 })
-export class IgxTimePickerComponent implements ControlValueAccessor, EditorProvider, OnInit, OnDestroy, DoCheck, AfterViewInit {
+export class IgxTimePickerComponent implements
+    IgxTimePickerBase,
+    ControlValueAccessor,
+    EditorProvider,
+    OnInit,
+    OnDestroy,
+    DoCheck,
+    AfterViewInit {
 
     private _value: Date;
