+++ projects/igniteui-angular/src/lib/time-picker/time-picker.directives.ts
@@ -16,7 +16,7 @@ import {
     Output,
     TemplateRef
 } from '@angular/core';
-import { IgxTimePickerComponent } from './time-picker.component';
+import { IGX_TIME_PICKER_COMPONENT, IgxTimePickerBase } from './time-picker.common';
 
 /** @hidden */
 @Directive({
@@ -29,8 +29,8 @@ export class IgxItemListDirective {
 
     public isActive: boolean;
 
-    constructor(@Host() @Inject(forwardRef(() => IgxTimePickerComponent))
-                public timePicker: IgxTimePickerComponent,
+    constructor(@Inject(IGX_TIME_PICKER_COMPONENT)
+                public timePicker: IgxTimePickerBase,
                 private elementRef: ElementRef) {}
 
     @HostBinding('attr.tabindex')
@@ -238,8 +238,8 @@ export class IgxHourItemDirective {
         return this.timePicker.selectedHour === this.value;
     }
 
-    constructor(@Host() @Inject(forwardRef(() => IgxTimePickerComponent))
-                public timePicker: IgxTimePickerComponent,
+    constructor(@Inject(IGX_TIME_PICKER_COMPONENT)
+                public timePicker: IgxTimePickerBase,
                 private itemList: IgxItemListDirective) {}
 
     @HostListener('click', ['value'])
@@ -280,8 +280,8 @@ export class IgxMinuteItemDirective {
         return this.timePicker.selectedMinute === this.value;
     }
 
-    constructor(@Host() @Inject(forwardRef(() => IgxTimePickerComponent))
-                public timePicker: IgxTimePickerComponent,
+    constructor(@Inject(IGX_TIME_PICKER_COMPONENT)
+                public timePicker: IgxTimePickerBase,
                 private itemList: IgxItemListDirective) {}
 
     @HostListener('click', ['value'])
@@ -322,8 +322,8 @@ export class IgxAmPmItemDirective {
         return this.timePicker.selectedAmPm === this.value;
     }
 
-    constructor(@Host() @Inject(forwardRef(() => IgxTimePickerComponent))
-                public timePicker: IgxTimePickerComponent,
+    constructor(@Inject(IGX_TIME_PICKER_COMPONENT)
+                public timePicker: IgxTimePickerBase,
                 private itemList: IgxItemListDirective) {}
 
     @HostListener('click', ['value'])
